move mp/score editing to matchup and support nonqualis
